+++ java/src/io/org/shypl/biser/io/SafeWriterProxy.java
@@ -42,7 +42,7 @@ public class SafeWriterProxy implements BiserWriter {
 	}
 
 	@Override
-	public void writeUint(int value) {
+	public void writeUint(long value) {
 		try {
 			writer.writeUint(value);
 		}
